+ Level now has compounds for collision detection.
+ Level.collision.compounds (for tile collision checks)
It compounds all of the collidable tiles together into the same object.
For all tiles in the compound, it references the compound object.
To use this compound with intended results, do not re-use any references
of the same memory.
(Data is contained in all tiles so the information may be accessed
anywhere, but can be filtered to only be checked once)